+++ app/src/app/components/contents/download/Download.tsx
@@ -1,40 +1,23 @@
-import { DownloadCloud } from "lucide-react";
 import {useEffect, useState} from "react";
 import { useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
-import appConfig from "../../../../config/appConfig";
-import { bytesToSize } from "../../../../helpers";
 import { FileMetadata, RootState, useAppDispatch } from "../../../../types";
 import { actions } from "../../../store/packets/packetThunks";
 import { PacketHeader } from "../packets";
-import {Button} from "../../Base/Button";
-import {download} from "../../../../lib/download";
+import DownloadButton from "./DownloadButton";
 
 export default function Download() {
   const { packet } = useSelector((state: RootState) => state.packets);
-  const { isAuthenticated } = useSelector((state: RootState) => state.login);
 
   const dispatch = useAppDispatch();
 
   const { packetId } = useParams();
-  const [error, setError] = useState("");
 
   useEffect(() => {
     if (packetId) {
       dispatch(actions.fetchPacketById(packetId));
     }
   }, [packetId]);
-
-  const downloadFile = (file: FileMetadata) => {
-    const url = `${appConfig.apiUrl()}/packets/file/${file.hash}?filename=${file.path}`;
-    download(url, file.path, isAuthenticated)
-        .catch(e =>  {
-          console.log("ERROR")
-          console.log(JSON.stringify(e.message))
-          setError(e.message);
-        })
-        .then(() => setError(""));
-  };
 
   if (Object.keys(packet).length === 0) {
     return <div>Loading...</div>;
@@ -49,14 +32,7 @@ export default function Download() {
             <div className="card custom-card">
               <div className="card-header">Download {data.path}</div>
               <div className="card-body">
-                <Button onClick={() => downloadFile(data)} variant="outline">
-                  {data.path}
-                  <span className="sidebar-icon p-2">
-                    <DownloadCloud />
-                  </span>
-                </Button>
-                <span className="small p-2 text-muted">({bytesToSize(data.size)})</span>
-                <div className="text-red-500">{error}</div>
+                <DownloadButton file={data}></DownloadButton>
               </div>
             </div>
           </li>
